requirements.txt
requests
pandas
python-dotenv
openpyxl
blockcypher
dash
plotly
python-binance